+++ VulkanGraphicsAPI/GridSystem.hpp
@@ -9,13 +9,13 @@ namespace SolEngine::DOD
     {
     public:
         GridSystem() = delete;
-        GridSystem(const glm::vec3 &dimensions);
+        GridSystem(const glm::uvec3 &dimensions);
         GridSystem(const size_t scalarDimensions);
 
     private:
         bool IsWithinAxisNodeCountLimit(const size_t count) const { return !(count > MAX_AXIS_NODE_COUNT); }
 
-        glm::vec3 _dimensions{ 0.f };
+        glm::uvec3 _dimensions{ 0.f };
         GridNodes _gridNodes{};
     };
 }
